Implement multiplication functionality for circular linked list
This commit adds a new functionality to a C program that manipulates numbers in a circular singly linked list. The program now includes a function that calculates and prints the multiplication of pairs of nodes in the list. If there is an odd number of nodes, the remaining node is multiplied by itself. If the list has less than 2 nodes, an error message is displayed.

Function prototype: void multiplicacao(Node *head, Node *tail);