Implement functionality to print elements preceded by a sequence of two odd numbers in a circular doubly linked list
This commit adds a new functionality to a C program that manipulates numbers in a circular doubly linked list. The program now includes a function to print all elements that are preceded by a sequence of two odd numbers. If an element does not have two preceding nodes, it checks if the two succeeding nodes are odd before printing the element. If the list has less than 3 nodes, it displays the message "Tamanho da lista é insuficiente :(" (List size is insufficient :().

Function prototype: void sequencia(Node *head, Node *tail);